removed whitespace removal in formatCert() in Utils.php
Had to revert the addition of the removal of the white spaces " ", leaving with just the additional \t removal because of how this function is used with header = true / false. If a header is presented (for example, this function is used to generate the metadata for consumption by the IdP) then the ----BEGIN/END CERTIFICATE---- has the spaces removed which breaks OpenSSL again and gives an invalid metadata error.

Leaving with \t removed fixes the problems we were having with tab indentation in the xml metadata held by the federation, however in any cases where the certs are indented with whitespaces a similar bug to the one I've uncovered could still occur. I'd suggest that the better was to do this would be in header = true, remove the ---BEGIN/END lines, flatten the certificate by removing \t and " " in addition to what's originally being removed then adding the header / footer back in again (as you're doing in headers = false). This would guarantee a working cert and not cause validation problems due to clobbering the headers.